+++ system/src/Grav/Framework/Flex/FlexCollection.php
@@ -50,6 +50,8 @@ class FlexCollection extends ObjectCollection implements FlexCollectionInterface
             'getTypePrefix' => true,
             'getType' => true,
             'getFlexDirectory' => true,
+            'hasFlexFeature' => true,
+            'getFlexFeatures' => true,
             'getCacheKey' => true,
             'getCacheChecksum' => true,
             'getTimestamp' => true,
@@ -95,16 +97,25 @@ class FlexCollection extends ObjectCollection implements FlexCollectionInterface
      * {@inheritdoc}
      * @see FlexCommonInterface::hasFlexFeature()
      */
-    public function hasFlexFeature($name): bool
+    public function hasFlexFeature(string $name): bool
+    {
+        return in_array($name, $this->getFlexFeatures(), true);
+    }
+
+    /**
+     * {@inheritdoc}
+     * @see FlexCommonInterface::hasFlexFeature()
+     */
+    public function getFlexFeatures(): array
     {
         $implements = class_implements($this);
         $list = [];
         foreach ($implements as $interface) {
-            $key = Inflector::hyphenize(preg_replace('/(.*\\\\)(.*?)Interface$/', '\\2', $interface));
-            $list[$key] = true;
+            $feature = Inflector::hyphenize(preg_replace('/(.*\\\\)(.*?)Interface$/', '\\2', $interface));
+            $list[] = $feature;
         }
 
-        return is_array($name) ? array_intersect_key($list, array_flip($name)) : $list[$name] ?? false;
+        return $list;
     }
 
     /**