+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added a new `bin/grav server` CLI command to easily run Symfony or PHP built-in webservers
     * Added `hasFlexFeature()` method to test if `FlexObject` or `FlexCollection` implements a given feature
+    * Added `getFlexFeatures()` method to return all features that `FlexObject` or `FlexCollection` implements
 1. [](#improved)
     * Better support for Symfony local server `symfony server:start`
     * Make `Route` objects immutable